+++ src/provider/AppContext.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { AppContextType } from "../types/types";
+
+const AppContext = createContext<AppContextType | undefined>(undefined);
+
+export default AppContext;
